Bug 1888504 - Add a dry-run mode for Bounce Tracking Protection. r=bvandersloot,anti-tracking-reviewers
Adds a mode where if bounce tracking protection is enabled it classifies
but does not purge trackers. This mode is helpful for testing the feature
without risking data loss. Telemetry is still collected normally.

Differential Revision: https://phabricator.services.mozilla.com/D206518

UltraBlame original commit: cb5c455b709b6dc796834ad27fb8c0b77fd5bd0e

+++ toolkit/components/antitracking/bouncetrackingprotection/BounceTrackingProtection.cpp
@@ -587,11 +587,17 @@ nsresult BounceTrackingProtection::PurgeBounceTrackersForStateGlobal(
             ("%s: Purge state for host: %s", __FUNCTION__,
              PromiseFlatCString(host).get()));
 
-    
-    rv = clearDataService->DeleteDataFromBaseDomain(host, false,
-                                                    TRACKER_PURGE_FLAGS, cb);
-    if (NS_WARN_IF(NS_FAILED(rv))) {
-      clearPromise->Reject(0, __func__);
+    if (StaticPrefs::privacy_bounceTrackingProtection_enableDryRunMode()) {
+      
+      
+      clearPromise->Resolve(host, __func__);
+    } else {
+      
+      rv = clearDataService->DeleteDataFromBaseDomain(host, false,
+                                                      TRACKER_PURGE_FLAGS, cb);
+      if (NS_WARN_IF(NS_FAILED(rv))) {
+        clearPromise->Reject(0, __func__);
+      }
     }
 
     aClearPromises.AppendElement(clearPromise);
